[feat] 랭킹 순위 뱃지 구현
