+++ feature/map/src/main/kotlin/com/unifest/android/feature/map/MapScreen.kt
@@ -2,28 +2,40 @@ package com.unifest.android.feature.map
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.PagerState
 import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -37,14 +49,17 @@ import com.naver.maps.map.compose.rememberCameraPositionState
 import com.naver.maps.map.overlay.OverlayImage
 import com.unifest.android.core.designsystem.ComponentPreview
 import com.unifest.android.core.designsystem.R
+import com.unifest.android.core.designsystem.component.NetworkImage
 import com.unifest.android.core.designsystem.component.SearchTextField
 import com.unifest.android.core.designsystem.component.TopAppBarNavigationType
 import com.unifest.android.core.designsystem.component.UnifestTopAppBar
+import com.unifest.android.core.designsystem.theme.Content2
+import com.unifest.android.core.designsystem.theme.Title2
+import com.unifest.android.core.designsystem.theme.Title5
 import com.unifest.android.core.designsystem.theme.UnifestTheme
 import com.unifest.android.core.domain.entity.BoothDetailEntity
 import com.unifest.android.core.domain.entity.BoothSpot
 import com.unifest.android.core.ui.DevicePreview
-import com.unifest.android.core.ui.component.BoothCard
 import com.unifest.android.core.ui.component.BoothFilterChips
 import com.unifest.android.core.ui.component.FestivalSearchBottomSheet
 import com.unifest.android.feature.map.viewmodel.MapUiState
@@ -141,6 +156,7 @@ internal fun MapScreen(
                 pagerState = pagerState,
                 boothList = uiState.boothList,
                 onNavigateToBooth = onNavigateToBooth,
+                isPopularMode = uiState.isPopularMode,
                 modifier = Modifier
                     .align(Alignment.BottomCenter)
                     .wrapContentHeight()
@@ -218,6 +234,7 @@ fun BoothCards(
     pagerState: PagerState,
     boothList: ImmutableList<BoothDetailEntity>,
     onNavigateToBooth: (Long) -> Unit,
+    isPopularMode: Boolean,
     modifier: Modifier = Modifier,
 ) {
     HorizontalPager(
@@ -228,9 +245,94 @@ fun BoothCards(
         BoothCard(
             boothInfo = boothList[page],
             onNavigateToBooth = onNavigateToBooth,
+            isPopularMode = isPopularMode,
+            ranking = page + 1,
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(horizontal = 12.dp),
+        )
+    }
+}
+
+@Composable
+fun BoothCard(
+    boothInfo: BoothDetailEntity,
+    onNavigateToBooth: (Long) -> Unit,
+    isPopularMode: Boolean,
+    ranking: Int,
+    modifier: Modifier = Modifier,
+) {
+    Card(
+        modifier = modifier.clickable {
+            onNavigateToBooth(boothInfo.id)
+        },
+        colors = CardDefaults.cardColors(containerColor = Color.White),
+        shape = RoundedCornerShape(12.dp),
+        elevation = CardDefaults.cardElevation(defaultElevation = 0.dp),
+    ) {
+        Box {
+            Row(
+                modifier = Modifier.padding(15.dp),
+            ) {
+                NetworkImage(
+                    imageUrl = "https://picsum.photos/86",
+                    modifier = Modifier
+                        .size(86.dp)
+                        .clip(RoundedCornerShape(16.dp)),
+                )
+                Column(
+                    modifier = Modifier.padding(start = 15.dp),
+                ) {
+                    Text(
+                        text = boothInfo.name,
+                        style = Title2,
+                    )
+                    Spacer(modifier = Modifier.height(3.dp))
+                    Text(
+                        text = boothInfo.description,
+                        maxLines = 2,
+                        overflow = TextOverflow.Ellipsis,
+                        style = Content2,
+                    )
+                    Spacer(modifier = Modifier.height(3.dp))
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        Icon(
+                            imageVector = ImageVector.vectorResource(R.drawable.ic_location_green),
+                            contentDescription = "Location Icon",
+                            tint = Color.Unspecified,
+                        )
+                        Spacer(modifier = Modifier.width(4.dp))
+                        Text(
+                            text = boothInfo.location,
+                            style = Title5,
+                        )
+                    }
+                }
+            }
+            if (isPopularMode) {
+                RankingBadge(ranking = ranking)
+            }
+        }
+    }
+}
+
+@Composable
+fun RankingBadge(ranking: Int) {
+    Box(
+        modifier = Modifier
+            .size(36.dp)
+            .clip(CircleShape)
+            .padding(start = 7.dp, top = 9.dp)
+            .background(Color(0xFFF5687E), CircleShape),
+        contentAlignment = Alignment.TopStart,
+    ) {
+        Text(
+            text = "${ranking}위",
+            color = Color.White,
+            style = Title5,
+            modifier = Modifier.align(Alignment.Center),
         )
     }
 }
@@ -320,7 +422,39 @@ fun BoothCardsPreview() {
             pagerState = rememberPagerState(pageCount = { boothList.size }),
             boothList = boothList.toImmutableList(),
             onNavigateToBooth = {},
+            isPopularMode = false,
             modifier = Modifier.height(116.dp),
         )
+    }
+}
+
+@ComponentPreview
+@Composable
+fun BoothCardPreview() {
+    UnifestTheme {
+        BoothCard(
+            boothInfo = BoothDetailEntity(
+                id = 1L,
+                name = "컴공 주점",
+                category = "",
+                description = "저희 주점은 일본 이자카야를 모티브로 만든 컴공인을 위한 주점입니다. 100번째 방문자에게 깜짝 선물 증정 이벤트를 하고 있으니 많은 관심 부탁드려요~!",
+                warning = "",
+                location = "청심대 앞",
+                latitude = 0f,
+                longitude = 0f,
+                menus = emptyList(),
+            ),
+            onNavigateToBooth = {},
+            isPopularMode = true,
+            ranking = 1,
+        )
+    }
+}
+
+@ComponentPreview
+@Composable
+fun RankingBadgePreview() {
+    UnifestTheme {
+        RankingBadge(ranking = 1)
     }
 }
